A few minor cleanups after controls anywhere PR
