+++ src/platform/plugins/shared/discover/common/app_locator.ts
@@ -136,7 +136,7 @@ export type DiscoverAppLocator = LocatorPublic<DiscoverAppLocatorParams>;
  */
 export interface MainHistoryLocationState {
   dataViewSpec?: DataViewSpec;
-  esqlControls?: ControlPanelsState<ESQLControlState> & SerializableRecord;
+  esqlControls?: ControlPanelsState<ESQLControlState>;
   isAlertResults?: boolean;
 }
