+++ src/platform/plugins/shared/discover/public/plugin_imports/embeddable_services.ts
@@ -8,13 +8,7 @@
  */
 
 export { ViewSavedSearchAction } from '../embeddable/actions/view_saved_search_action';
+export { addPanelFromLibrary } from '../embeddable/utils/add_panel_from_library';
 export { getSearchEmbeddableFactory } from '../embeddable/get_search_embeddable_factory';
 export { getLegacyLogStreamEmbeddableFactory } from '../embeddable/get_legacy_log_stream_embeddable_factory';
 export { getSearchEmbeddableTransforms } from '../../common/embeddable';
-export { addControlsFromSavedSession } from '../embeddable/utils/add_controls_from_saved_session';
-export {
-  SAVED_OBJECT_REF_NAME,
-  apiPublishesEditablePauseFetch,
-  apiHasUniqueId,
-} from '@kbn/presentation-publishing';
-export { apiPublishesESQLVariables } from '@kbn/esql-types';
